Implement CubQ.dequeue_ack/2 and CubDB.pop_ack/2
This is useful when implementing "at least once" semantics with multiple
consumers of a queue.

Basically, `CubQ.dequeue_ack(pid, timeout)` (or the symmetrical
`CubQ.pop_ack(pid, timeout)`) dequeues an item, but expects the consumer
to explicitly call `ack/2` or `nack/2` to confirm that the item was
successfully consumed or not, before the `timeout` elapses.

If `nack/2` is called, or if the timeout elapses before `ack/2` is
called, the item is put back in the queue, so it can be dequeued again
by a consumer. This is done in a durable and atomic way: even if the
`CubQ` process exits, when restarted it will still keep track of the
pending acks, and put the items back to the queue if not acked on time.